feat: dump yaml including signature for flex flow run snapshot
