+++ src/promptflow-azure/promptflow/azure/_entities/_flow.py
@@ -14,7 +14,7 @@ from promptflow._constants import FlowLanguage
 from promptflow._proxy import ProxyFactory
 from promptflow._sdk._constants import SERVICE_FLOW_TYPE_2_CLIENT_FLOW_TYPE, AzureFlowSource, FlowType
 from promptflow._sdk._utils import PromptflowIgnoreFile, load_yaml, remove_empty_element_from_dict
-from promptflow._utils.flow_utils import dump_flow_dag, load_flow_dag, resolve_flow_path
+from promptflow._utils.flow_utils import dump_flow_dag, load_flow_dag, resolve_flow_path, update_signatures
 from promptflow._utils.logger_utils import LoggerFactory
 from promptflow.azure._ml import AdditionalIncludesMixin, Code
 
@@ -132,14 +132,6 @@ class Flow(AdditionalIncludesMixin):
         flow_dag.pop(ADDITIONAL_INCLUDES, None)
         return True
 
-    @classmethod
-    def _resolve_signature(cls, code: Path, data: dict):
-        """Resolve signature for flex flow. Return True if resolved."""
-        from promptflow.client import PFClient
-
-        pf = PFClient()
-        return pf.flows._update_signatures(code=code, data=data)
-
     # region AdditionalIncludesMixin
     @contextmanager
     def _try_build_local_code(self) -> Optional[Code]:
@@ -168,7 +160,7 @@ class Flow(AdditionalIncludesMixin):
                 ProxyFactory().create_inspector_proxy(self.language).prepare_metadata(
                     flow_file=flow_directory / flow_file, working_dir=flow_directory
                 )
-                dag_updated = self._resolve_signature(flow_dir, flow_dag) or dag_updated
+                dag_updated = update_signatures(code=flow_dir, data=flow_dag) or dag_updated
                 self._environment = self._resolve_environment(flow_dir, flow_dag)
                 if dag_updated:
                     dump_flow_dag(flow_dag, flow_dir)
