+++ src/promptflow-devkit/promptflow/_sdk/operations/_local_storage_operations.py
@@ -37,7 +37,7 @@ from promptflow._sdk._utils import (
 from promptflow._sdk.entities import Run
 from promptflow._sdk.entities._flows import Flow
 from promptflow._utils.exception_utils import PromptflowExceptionPresenter
-from promptflow._utils.flow_utils import is_prompty_flow
+from promptflow._utils.flow_utils import dump_flow_dag, is_prompty_flow, update_signatures
 from promptflow._utils.logger_utils import LogContext, get_cli_sdk_logger
 from promptflow._utils.multimedia_utils import MultimediaProcessor
 from promptflow._utils.utils import prepare_folder
@@ -255,8 +255,13 @@ class LocalStorageOperations(AbstractBatchRunStorage):
             ignore=shutil.ignore_patterns(*patterns),
             dirs_exist_ok=True,
         )
-        # replace DAG file with the overwrite one
-        if not self._eager_mode and not self._is_prompty_flow:
+        if self._eager_mode:
+            yaml_dict = copy.deepcopy(flow._data)
+            update_signatures(code=flow.code, data=yaml_dict)
+            # for eager mode, we need to update signature for it
+            dump_flow_dag(flow_dag=yaml_dict, flow_path=self._dag_path)
+        elif not self._is_prompty_flow:
+            # replace DAG file with the overwrite one
             self._dag_path.unlink()
             shutil.copy(flow.path, self._dag_path)
